feat: Add createUser Firebase Cloud Function
This commit adds a new Firebase Cloud Function called "createUser". The function is triggered when a new user is created in the authentication system. It retrieves the user's information and stores it in the Firestore database under the "users" collection. Additionally, it logs a message with the user's information using the Firebase Functions logger.