feat: Add file upload functionality to navbar
This commit adds the necessary code to enable file upload functionality to the navbar component. It includes the implementation of the handleFileChange and handleUpload functions, which allow users to select a video file and upload it to the server using the uploadVideo function from the firebase/functions module. The uploaded file is then processed by the server and a success or error message is displayed to the user.